+++ crates/bevy_pbr/src/render/mesh.rs
@@ -1,4 +1,4 @@
-use crate::{material_bind_groups::MaterialBindGroupIndex, render::mesh_bindings::MeshLayoutsBuilders as _};
+use crate::{material_bind_groups::MaterialBindGroupIndex, render::mesh_bindings::MeshLayoutsBuilders as _, skin::skin_uniforms_from_world};
 pub use bevy_render::mesh::render::*;
 use bevy_asset::{embedded_asset, load_embedded_asset, AssetId};
 use bevy_camera::{
@@ -200,7 +200,7 @@ impl Plugin for MeshRenderPlugin {
                 .init_resource::<ViewKeyCache>()
                 .init_resource::<ViewSpecializationTicks>()
                 .init_resource::<GpuPreprocessingSupport>()
-                .init_resource::<SkinUniforms>()
+                .add_systems(RenderStartup, skin_uniforms_from_world)
                 .add_systems(
                     Render,
                     check_views_need_specialization.in_set(PrepareAssets),
