+++ crates/bevy_render/src/mesh/mod.rs
@@ -1,6 +1,7 @@
 pub mod allocator;
 pub mod material_bind_group;
 pub mod render;
+pub mod skin;
 use crate::{
     render_asset::{PrepareAssetError, RenderAsset, RenderAssetPlugin, RenderAssets},
     texture::GpuImage,
